+++ src/__test__/Login-form.test.js
@@ -1,6 +1,6 @@
 import React from 'react';
-import { render, fireEvent } from '@testing-library/react';
-import LoginForm from '../components/Login-form';
+import { fireEvent, render } from '@testing-library/react';
+import LoginForm from '../components/Login/LoginForm';
 
 it('renders correctly', () => {
     const {queryByTestId} = render(<LoginForm />);
