+++ src/__test__/Login.test.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { render, fireEvent } from '@testing-library/react';
-import Login from '../components/Login';
+import Login from '../components/Login/Login';
 
 it('renders correctly', () => {
     const {queryByTestId} = render(<Login />);
