+++ src/__test__/Map.test.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { render } from '@testing-library/react';
-import Map from '../components/Map';
+import Map from '../components/Map/Map';
 
 it('renders correctly', () => {
     const {queryByTestId} = render(<Map />);
