--- src/components/Profile.js
@@ -1,5 +1,0 @@
-import React from 'react';
-
-const Profile = () => <h1>Profile</h1>;
-
-export default Profile;
