+++ src/components/Profile/Profile.js
@@ -1,0 +1,5 @@
+import React from 'react';
+
+const Profile = () => <h1 data-testid='Profile'>Profile</h1>;
+
+export default Profile;
